test(#20): 테스트 통과 기원
